Back to plotting in a separate file.
This reads an hdf file for the data and puts it in an xarray (pip
install xarray).  This facilitates multi-dimensional slicing by axis
name.  We then produce plots separately for:

- Time/dof for matvec and assembly (aij and matrix-free)
- Bytes/dof for assembled and matrix-free matrices

The dataset contains timings across process counts (or can do).  So we
can do scaling plots from this data later.